+++ frontend/src/components/inputs/ChatInput.tsx
@@ -17,6 +17,7 @@ export default function ChatInput({
 	const textareaRef = useRef<HTMLTextAreaElement>(null);
 	const [isRecording, setIsRecording] = useState(false);
 	const { loading } = useAppContext();
+	const { isLikelyMobile } = useAppHook();
 	const {
 		query,
 		abortQuery,
@@ -31,8 +32,6 @@ export default function ChatInput({
 		setPreviewImage,
 		handleSubmit,
 	} = useChatContext();
-
-	const { isMobile } = useAppHook();
 
 	// Initialize the recorder controls using the hook
 	const recorderControls = useVoiceVisualizer();
@@ -89,7 +88,8 @@ export default function ChatInput({
 						query.length > 0
 					) {
 						e.preventDefault();
-						if (!loading && !isMobile()) handleSubmit(query, images);
+						if (!loading && !isLikelyMobile())
+							handleSubmit(query, images);
 					}
 				}}
 			/>