+++ frontend/src/hooks/useAgent.ts
@@ -1,8 +1,8 @@
 import agentService, { Agent } from "@/lib/services/agentService";
 import { useEffect, useState } from "react";
 import ToolConfig from "@/lib/config/tool";
-import useModel from "./useModel";
 import { DropdownMenuCheckboxItemProps } from "@radix-ui/react-dropdown-menu";
+import useModel from "./useModel";
 
 type Checked = DropdownMenuCheckboxItemProps["checked"];
 
@@ -31,7 +31,9 @@ export const INIT_AGENT_STATE: AgentState = {
 };
 
 export function useAgent() {
-	const { model } = useModel();
+	const { model, useModelsEffect } = useModel();
+	useModelsEffect();
+
 	const [agent, setAgent] = useState<Agent>(INIT_AGENT_STATE.agent);
 	const [agents, setAgents] = useState<Agent[]>([]);
 	const [webSearchCheck, setWebSearchCheck] = useState<Checked>(() => {
@@ -56,13 +58,14 @@ export function useAgent() {
 	useEffect(() => {
 		setAgent({
 			...agent,
+			model: model ?? "",
 			presidio: {
 				analyze: piiAnalyzeCheck ? true : false,
 				anonymize: piiAnonymizeCheck ? true : false,
 				redact: false,
 			},
 		});
-	}, [piiAnalyzeCheck, piiAnonymizeCheck]);
+	}, [piiAnalyzeCheck, piiAnonymizeCheck, model]);
 
 	const setAgentSystemMessage = (system: string) => {
 		setAgent({ ...agent, prompt: system });